Rename rule-executors to dom-actions
